backend/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
pyjwt==2.8.0
django-cors-headers==4.3.1
setuptools==68.2.2